Bluetooth: inline BluetoothUtils into Handler classes.
